backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.5,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
boto3>=1.34,<2.0
neo4j>=5.18,<6.0
ddtrace>=2.8,<3.0
python-multipart>=0.0.9,<1.0
